Feature/vea 10 authentication and authorization (#2)
* VEA-11 Added security dependencies.

Security dependencies include spring boot starter security, and JWT
packages (from JJWT).

* VEA-11 Added configuration properties.

The properties include the token duration, and secret properties. The
secret properties include the token secret key.

The secret properties file was added to gitignore so that the secrets
are not tracked by Git, and an example file was created to show the
format of the secret properties file.

* VEA-11 Util class for checkPasswordValidity.

The checkPasswordValidity method was moved to a new class (AuthUtils) as
a static method.

* VEA-11 Util class for JWT.

Added a utils class for JSON web tokens. This class contains methods to
generate a token and to extract claims from a token.

* VEA-11 Customer details service.

Added implementations for UserDetails and UserDetailsService
(CustomerDetails and CustomerDetailsService). This will be used in
authentication.

* VEA-11 Filter for JWT authentication.

Added a filter for JWT authentication. This filter checks for the JWT,
checks if it's valid, and authorizes the user.

* VEA-11 Config for security and authentication.

Added configuration for authentication (authentication provider,
authentication manager, password encoder) and security (using the JWT
auth filter).

* VEA-11 Register and login endpoints.

Added endpoints for register and login. This includes controller and
service, as well as the request and response objects.

* VEA-11 Removed confirm password from customer.

Removed the transient confirm password field from customer since it is
handled by the RegisterRequest class. Also added more constructors.

* VEA-11 Removed auth from customer package.

Removed auth methods and endpoints from the customer package since it is
now handled separately by the auth controller.

* VEA-11 Added responses in auth filter.

Added unauthorized and bad request responses for the JWT auth filter
(instead of just continuing to the next filter).

* VEA-11 Used JWT library exceptions.

The JWT library being used (jjwt) has its own exceptions to use when
parsing a JWT. These are now being used in the JWT auth filter.

* VEA-11 Made constant for whitelisted routes.

Added class for auth constants with a constant for the whitelisted
routes (register and login) and used this constant in the security
filter chain.

* VEA-11 Made the util classes final.

Made AuthUtils and JwtUtils final and gave them a private constructor so
that they cannot be instantiated. This way, they can only be used by
their static properties and methods.

* VEA-11 Auth filter ignores whitelisted routes.

For the whitelisted routes (register and login), the JWT auth filter
does do any authorization actions or processing.

* VEA-11 Helper methods setting header in filter.

Extracted logic to set the status and WWW-Authenticate header from the
JWT auth filter to its own private helper methods.

* VEA-11 Exception handler for authentication fail.

Added an exception handler for failed authentication.

* VEA-11 Added JavaDocs.

* VEA-12 Fixed integration test errors.

Due to the new auth functionality, the previous tests have errors. The
errors are now fixed (although the tests fail).

* VEA-12 Manually added servlet path in tests.

For some reason, Spring tests do not populate the servlet path in the
request, which causes errors in the JWT auth filter. Created a request
post processor to add the servlet path manually.

* VEA-12 Reorganized tests.

Moved the auth-related tests to their own test package, and kept one
customer-related test for the customer package.

* VEA-12 Made test utils class.

This class contains helper methods for tests.

* VEA-12 Refactored tests to be less error-prone.

Used variables for the URL path so that it doesn't rely on string
literals.

Also improved readability.

* VEA-12 Added tests for login endpoint.