Feature/vea 7 get user details (#3)
* VEA-13 Changed customer username to id.

Previously, the customer's email was used as their username. Now, it is
their id.

* VEA-13 Get customer endpoint now uses principal.

Previously, the get customer endpoint used a path variable containing
the id of the customer. Now, it uses the principal (which is populated
from the auth token in the JWT auth filter).

* VEA-13 Customer updatedAt and lastLogin fields.

Added updatedAt and lastLogin fields to customer.

* VEA-13 Logging in changes lastLogin field.

When a customer logs in, their lastLogin field is now updated.

* VEA-13 Customer response excludes id and password.

Response from customer API no longer includes the customer's id or
(hashed) password.

* VEA-14 Updated customer API test.

Instead of using the id in the ger customer request, just the JWT is
used. This corresponds with the get customer API change.

* VEA-14 Ensured customer id and passwords null.

Added functionality to the get customer test to ensure the returned
customer has null id and password.