Feature/vea 5 edit customer (#4)
* VEA-15 Update customer endpoint.

Added endpoint for updating customer. Currently only updates name since
other updatable fields require more steps to update.

* VEA-15 Added update password endpoint.

* VEA-15 Error response viewable to anonymous users.

Updated security config to allow error responses for anonymous users.

* VEA-16 Added DTO for customer update.

Instead of using a map for the request body when updating a customer, a
DTO is used instead for simpler validation of the request body.

* VEA-16 JSON validation fail on unknown properties.

Added property to make JSON deserialization/validation fail when unknown
properties are encountered. This means request bodies cannot have
additional properties than what is specified in the request body's
class, for example.

* VEA-16 Added tests for update customer endpoint.

* VEA-16 Added DTO for customer password update.

Instead of using a map for the request body when updating a customer's
password, a DTO is used instead for simpler validation of the request
body.

* VEA-16 Added tests for update password endpoint.

* VEA-16 New password cannot equal old password.

Added constraint in update customer password endpoint where the new
password cannot be the same as the old password.

* VEA-16 Test case for different new password.

Added test case to test the requirement of the new password not being
the same as the old password.

* VEA-16 Moved auth DTOs into their own package.

Made a package for auth DTOs (LoginRequest, RegisterRequest,
AuthResponse) and moved them into the package for better organization.

* VEA-16 Made customers unique between tests.

Made the names and emails of customers unique between test classes.
There were some test failures because the customers created from one
test class were still in the database for the next test class. This
change makes sure that the customers are completely unique for every
test.

* VEA-16 Cleaned up TestUtils class.

Fixed a spelling error and removed extractClaimsFromToken method because
it is no longer being used anywhere.

* VEA-16 Added Autowired to customer controller.

Forgot to add @Autowired annotation to CustomerController constructor,
so it is being added now. It is not necessary, but it is being added for
consistency.