Feature/vea 6 delete user (#5)
* VEA-18 Delete customer endpoint.

Added endpoint for deleting customer.

* VEA-19 Test case for delete customer endpoint.

Added test case for delete customer endpoint.

* VEA-19 Fixed test cases by adding auth token.

Some of the tests that expected a 400 response received that response
because they were missing the auth token. These changes add the auth
token so that the test cases are correctly testing the endpoints.

* VEA-19 Delete customer response status 204.

Changed the response status of the delete customer endpoint to 204 No
Content instead of 200 OK.

* VEA-19 Delete customer test response status 204.

Changed the delete customer test to expect a response status of 204.

* VEA-19 Added Postman requests to repository.

Added the Postman collection to test the endpoints of this API and the
environment for the variables used by Postman.