Feature/vea 23 initial deploy (#7)
* VEA-26 Added MySQL dependency.

* VEA-26 Added configuration for database.

* VEA-26 Added log config for production.

Added configuration for logging to a file for prod active profile.

* VEA-26 Added banner.txt.

* VEA-26 Increased version to 1.1.0-SNAPSHOT.

* VEA-26 Added token properties for test.

The tests were failing because these properties were missing. I thought
that the test also gets properties from application.properties in the
main folder, but I have to add them here in the test folder as well for
the tests to work. I don't know why the tests passed before without
these properties in the test folder.