Feature/vea 27 send and verify endpoints (#9)
* VEA-29 Added code entity class.

Added class representing a verification code entity, along with a
builder class creating the code entity.

* VEA-29 Endpoint for sending a verification code.

* VEA-30 Fixed logical error in Code.isActive.

The codeIsExpired local variable had the wrong value because I used the
wrong logical operator (< instead of >=). Fixed this by changing the
operator.

* VEA-30 Code.isActive checks if already fulfilled.

Added a check in the Code.isActive method to make sure the code has not
already been fulfilled.

* VEA-30 Increment incorrect attempts for code.

Added a method for incrementing the number of incorrect attempts for the
code entity.

* VEA-30 Endpoint for verifying a verification code.

* VEA-31 Refactored send and verify code endpoints.

Added a consistent response type for both the send and verify code
endpoints and improved error handling. Added a code utils component
instead of static utils class to be more consistent with the Spring
framework.

* VEA-31 Tests for send and verify code endpoints.

* VEA-31 Made TestUtils use generic methods.

Refactored TestUtils so that it uses generic methods instead of a method
for each type.

* VEA-31 Fixed name of CodeIntegrationTest.