Feature/vea 28 send email functionality (#10)
* VEA-34 Emailer with API emailer implementation.

Created emailer interface and an API emailer implementation that calls
an API to send the email. This also includes the configuration,
exceptions, and DTOs.

* VEA-34 Updated pom.xml to 1.2.0-SNAPSHOT.

* VEA-34 Improved send email error handling.

Added error handlers in CodeController to handle errors from sending
email via ApiEmailer and added documentation of the errors to expect in
the ApiEmailer method Javadocs.

* VEA-34 Updated secrets.properties.example.

Updated secrets.properties.example with new properties for the API
emailer.

* VEA-34 Renamed code.exceptions package.

Renamed package com.danieldigiovanni.email.code.exceptions to
com.danieldigiovanni.email.code.exception.

* VEA-35 Added Spring boot mailer dependency.

Added dependency spring-boot-starter-mail dependency for sending email
to Mailtrap.

* VEA-35 Mailtrap emailer.

Created implementation of the Emailer interface for sending email to
Mailtrap.

* VEA-35 Refactored emailer bean config.

Instead of having an EmailerConfig that decides what implementation of
Emailer to initialize, separate configs for ApiEmailer and
MailtrapEmailer are now being used, with the Spring annotation
@ConditionalOnProperty being used to decide which bean to initialize.

* VEA-35 Updated secrets.properties.example.

* VEA-35 Added NoOpEmailer.

Added a NoOpEmailer implementation of Emailer interface that just logs a
message instead of sending any email out. Also updated test properties
to use this implementation of Emailer.

* VEA-35 Added customer field to code entity.

Added a field to the code entity representing the customer that created
the code. Also added a test case to test for this case.

* VEA-35 Added emailer config logs.

Added logs when initializing each of the Emailer implementations.

* VEA-35 Added more emailer logs.

Added logs in ApiEmailer and MailtrapEmailer to log events like sending
email and refreshing token.

* VEA-35 Updated Postman collections and environment