Feature/vea 43 consistent error handling (#16)
* VEA-44 Application-wide exception handling.

Moved exception handlers in controllers to exception handlers in an
application-wide controller advice component.

* VEA-44 Standard error response body.

Made a standard format for the response body of error responses.

* VEA-44 Unhandled exception handler.

Made a controller that handles all exceptions that did not get
handled (like the last line of defense in our application for errors).

* VEA-44 Documentation for the error package.

* VEA-44 Code style fix.

* VEA-44 Refactored package structure.

Made error and exception handlers its own package, made filters their
own package.

* VEA-44 Made incorrect code part of normal flow.

Instead of incorrect code being an exception, it is now part of the
normal program flow. This means it is handled in the service class
(CodeService) instead of in an exception handler.